# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Focus Play local API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/sources/search": {
      "get": {
        "summary": "Search",
        "operationId": "search_api_sources_search_get",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 3,
              "maxLength": 200,
              "title": "Q"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SearchResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/lessons": {
      "get": {
        "summary": "Lessons",
        "operationId": "lessons_api_lessons_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/Lesson"
                  },
                  "type": "array",
                  "title": "Response Lessons Api Lessons Get"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create",
        "operationId": "create_api_lessons_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LearningRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Lesson"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/lessons/{lid}": {
      "get": {
        "summary": "Snapshot",
        "operationId": "snapshot_api_lessons__lid__get",
        "parameters": [
          {
            "name": "lid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Lid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Lesson"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/lessons/{lid}/cancel": {
      "post": {
        "summary": "Cancel",
        "operationId": "cancel_api_lessons__lid__cancel_post",
        "parameters": [
          {
            "name": "lid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Lid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Lesson"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/lessons/{lid}/retry": {
      "post": {
        "summary": "Retry",
        "operationId": "retry_api_lessons__lid__retry_post",
        "parameters": [
          {
            "name": "lid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Lid"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Lesson"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/lessons/{lid}/shorts/{sid}/explanations": {
      "post": {
        "summary": "Explanation",
        "operationId": "explanation_api_lessons__lid__shorts__sid__explanations_post",
        "parameters": [
          {
            "name": "lid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Lid"
            }
          },
          {
            "name": "sid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Sid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExplanationRequest"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Lesson"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/lessons/{lid}/events": {
      "get": {
        "summary": "Events",
        "operationId": "events_api_lessons__lid__events_get",
        "parameters": [
          {
            "name": "lid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Lid"
            }
          },
          {
            "name": "after",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 0,
              "title": "After"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/assets/{asset_id}/metadata": {
      "get": {
        "summary": "Asset Metadata",
        "operationId": "asset_metadata_api_assets__asset_id__metadata_get",
        "parameters": [
          {
            "name": "asset_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Asset Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AssetRecord"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/assets/{asset_id}": {
      "get": {
        "summary": "Asset Content",
        "operationId": "asset_content_api_assets__asset_id__get",
        "parameters": [
          {
            "name": "asset_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Asset Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/audio/{name}": {
      "get": {
        "summary": "Audio",
        "operationId": "audio_api_audio__name__get",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Name"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "AcquisitionLedger": {
        "properties": {
          "round_limit": {
            "type": "integer",
            "maximum": 2.0,
            "minimum": 1.0,
            "title": "Round Limit",
            "default": 2
          },
          "transcript_limit": {
            "type": "integer",
            "maximum": 30.0,
            "minimum": 1.0,
            "title": "Transcript Limit",
            "default": 16
          },
          "per_round_limit": {
            "type": "integer",
            "maximum": 15.0,
            "minimum": 1.0,
            "title": "Per Round Limit",
            "default": 8
          },
          "rounds": {
            "items": {
              "$ref": "#/components/schemas/AcquisitionRound"
            },
            "type": "array",
            "maxItems": 2,
            "title": "Rounds",
            "default": []
          },
          "tried_video_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "maxItems": 30,
            "title": "Tried Video Ids",
            "default": []
          },
          "search_provider_calls": {
            "type": "integer",
            "title": "Search Provider Calls",
            "default": 0
          },
          "transcript_provider_calls": {
            "type": "integer",
            "title": "Transcript Provider Calls",
            "default": 0
          },
          "youtube_http_calls": {
            "type": "integer",
            "title": "Youtube Http Calls",
            "default": 0
          },
          "youtube_quota_units": {
            "type": "integer",
            "title": "Youtube Quota Units",
            "default": 0
          },
          "transcript_http_calls": {
            "type": "integer",
            "title": "Transcript Http Calls",
            "default": 0
          },
          "search_cache_hits": {
            "type": "integer",
            "title": "Search Cache Hits",
            "default": 0
          },
          "transcript_cache_hits": {
            "type": "integer",
            "title": "Transcript Cache Hits",
            "default": 0
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "AcquisitionLedger"
      },
...
      "AssetRecord": {
        "properties": {
          "id": {
            "type": "string",
            "pattern": "^[a-f0-9]{64}$",
            "title": "Id"
          },
          "content_hash": {
            "type": "string",
            "pattern": "^[a-f0-9]{64}$",
            "title": "Content Hash"
          },
          "kind": {
            "type": "string",
            "enum": [
              "image",
              "screenshot"
            ],
            "title": "Kind"
          },
          "mime_type": {
            "type": "string",
            "enum": [
              "image/png",
              "image/jpeg"
            ],
            "title": "Mime Type"
          },
          "byte_size": {
            "type": "integer",
            "maximum": 5000000.0,
            "minimum": 1.0,
            "title": "Byte Size"
          },
          "width": {
            "type": "integer",
            "maximum": 4096.0,
            "minimum": 1.0,
            "title": "Width"
          },
          "height": {
            "type": "integer",
            "maximum": 4096.0,
            "minimum": 1.0,
            "title": "Height"
          },
          "original_source": {
            "type": "string",
            "maxLength": 500,
            "minLength": 1,
            "title": "Original Source"
          },
          "creator": {
            "type": "string",
            "maxLength": 150,
            "minLength": 1,
            "title": "Creator"
          },
          "permission_basis": {
            "type": "string",
            "maxLength": 500,
            "minLength": 5,
            "title": "Permission Basis"
          },
          "attribution": {
            "type": "string",
            "maxLength": 300,
            "minLength": 5,
            "title": "Attribution"
          },
          "license_url": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 500
              },
              {
                "type": "null"
              }
            ],
            "title": "License Url"
          },
          "acquired_at": {
            "type": "number",
            "minimum": 0.0,
            "title": "Acquired At"
          },
          "managed_filename": {
            "type": "string",
            "pattern": "^[a-f0-9]{64}\\.(png|jpg)$",
            "title": "Managed Filename"
          },
          "source_context": {
            "type": "string",
            "maxLength": 300,
            "minLength": 5,
            "title": "Source Context"
          },
          "illustrative": {
            "type": "boolean",
            "title": "Illustrative"
          },
          "status": {
            "type": "string",
            "enum": [
              "validated",
              "ready",
              "missing",
              "failed"
            ],
            "title": "Status",
            "default": "validated"
          },
          "failure_reason": {
            "type": "string",
            "maxLength": 300,
            "title": "Failure Reason",
            "default": ""
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "content_hash",
          "kind",
          "mime_type",
          "byte_size",
          "width",
          "height",
          "original_source",
          "creator",
          "permission_basis",
          "attribution",
          "acquired_at",
          "managed_filename",
          "source_context",
          "illustrative"
        ],
        "title": "AssetRecord"
      },
...
      "ExplanationRequest": {
        "properties": {
          "kind": {
            "type": "string",
            "enum": [
              "example",
              "again"
            ],
            "title": "Kind"
          },
          "added_seconds": {
            "type": "integer",
            "const": 40,
            "title": "Added Seconds",
            "default": 40
          },
          "request_id": {
            "type": "string",
            "maxLength": 100,
            "minLength": 8,
            "title": "Request Id"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "kind",
          "request_id"
        ],
        "title": "ExplanationRequest"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "LearningRequest": {
        "properties": {
          "goal": {
            "type": "string",
            "maxLength": 500,
            "minLength": 3,
            "title": "Goal"
          },
          "prior_knowledge": {
            "type": "string",
            "maxLength": 500,
            "minLength": 1,
            "title": "Prior Knowledge",
            "default": "beginner"
          },
          "time_budget_seconds": {
            "type": "integer",
            "maximum": 1200.0,
            "minimum": 60.0,
            "title": "Time Budget Seconds",
            "default": 300
          },
          "language": {
            "type": "string",
            "const": "en",
            "title": "Language",
            "default": "en"
          },
          "request_id": {
            "type": "string",
            "maxLength": 100,
            "minLength": 8,
            "title": "Request Id"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "goal",
          "request_id"
        ],
        "title": "LearningRequest"
      },
      "Lesson": {
        "properties": {
          "readiness": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Readiness"
              },
              {
                "type": "null"
              }
            ]
          },
          "planning": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PlanningState"
              },
              {
                "type": "null"
              }
            ]
          },
          "acquisition": {
            "$ref": "#/components/schemas/AcquisitionLedger"
          },
          "duration_ledger": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/DurationLedger"
              },
              {
                "type": "null"
              }
            ]
          },
          "teaching_plan_version": {
            "type": "integer",
            "enum": [
              1,
              2
            ],
            "title": "Teaching Plan Version",
            "default": 1
          },
          "examples": {
            "items": {
              "$ref": "#/components/schemas/ExampleRecord"
            },
            "type": "array",
            "maxItems": 2,
            "title": "Examples",
            "default": []
          },
          "coverage_history": {
            "items": {
              "$ref": "#/components/schemas/CoverageEntry"
            },
            "type": "array",
            "maxItems": 83,
            "title": "Coverage History",
            "default": []
          },
          "plan_diagnostics": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "maxItems": 8,
            "title": "Plan Diagnostics",
            "default": []
          },
          "id": {
            "type": "string",
            "title": "Id"
          },
          "request": {
            "$ref": "#/components/schemas/SavedLearningRequest"
          },
          "objectives": {
            "items": {
              "$ref": "#/components/schemas/Objective"
            },
            "type": "array",
            "title": "Objectives",
            "default": []
          },
          "short_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Short Ids",
            "default": []
          },
          "shorts": {
            "items": {
              "$ref": "#/components/schemas/Short"
            },
            "type": "array",
            "title": "Shorts",
            "default": []
          },
          "sources": {
            "items": {
              "$ref": "#/components/schemas/Source"
            },
            "type": "array",
            "title": "Sources"
          },
          "status": {
            "type": "string",
            "enum": [
              "queued",
              "preparing",
              "partially_ready",
              "ready",
              "failed",
              "cancelled",
              "interrupted"
            ],
            "title": "Status",
            "default": "queued"
          },
          "planned_duration_ms": {
            "type": "integer",
            "title": "Planned Duration Ms",
            "default": 0
          },
          "original_planned_duration_ms": {
            "type": "integer",
            "title": "Original Planned Duration Ms",
            "default": 0
          },
          "extra_allowance_ms": {
            "type": "integer",
            "title": "Extra Allowance Ms",
            "default": 0
          },
          "job": {
            "$ref": "#/components/schemas/Job"
          },
          "provider_settings": {
            "additionalProperties": true,
            "type": "object",
            "title": "Provider Settings",
            "default": {}
          },
          "metrics": {
            "additionalProperties": {
              "type": "number"
            },
            "type": "object",
            "title": "Metrics",
            "default": {}
          },
          "video_rankings": {
            "items": {
              "$ref": "#/components/schemas/CandidateScore"
            },
            "type": "array",
            "title": "Video Rankings",
            "default": []
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "request",
          "sources",
          "job"
        ],
        "title": "Lesson"
      },
...
      "Readiness": {
        "properties": {
          "version": {
            "type": "integer",
            "const": 1,
            "title": "Version",
            "default": 1
          },
          "ready_short_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Ready Short Ids",
            "default": []
          },
          "missing_media_short_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Missing Media Short Ids",
            "default": []
          },
          "initial_contiguous_media_ms": {
            "type": "integer",
            "title": "Initial Contiguous Media Ms",
            "default": 0
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "Readiness"
      },
...
      "SearchItem": {
        "properties": {
          "video_id": {
            "type": "string",
            "title": "Video Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "channel": {
            "type": "string",
            "title": "Channel"
          },
          "url": {
            "type": "string",
            "title": "Url"
          },
          "duration_seconds": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Duration Seconds"
          },
          "transcript_status": {
            "type": "string",
            "const": "needed",
            "title": "Transcript Status",
            "default": "needed"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "video_id",
          "title",
          "channel",
          "url"
        ],
        "title": "SearchItem"
      },
      "SearchResponse": {
        "properties": {
          "results": {
            "items": {
              "$ref": "#/components/schemas/SearchItem"
            },
            "type": "array",
            "title": "Results"
          },
          "cached": {
            "type": "boolean",
            "title": "Cached"
          },
          "acquisition": {
            "type": "string",
            "const": "assisted",
            "title": "Acquisition",
            "default": "assisted"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "results",
          "cached"
        ],
        "title": "SearchResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}